Don't use flatMapLatest with network
Description:
It was canceling the previous lambda when a new value arrived.

This is problematic, because it can lead to two overlapping network requests mutating the same data, and we won't know for sure which one was acted on last.

Test Plan:
Open logcat. Do edits, including when the network is currently saving
Expected: It does not submit the next request until the previous one has success or failure
Expected: The final value is always committed, not left in a ! state